+++ Changes.md
@@ -2,6 +2,10 @@
 
 ### Added
 
+- Allow context to be a hashref or arrayref
+
+- Use JSON::MaybeXS instead of JSON
+
 - Added MooX::JSON_LD helper module
 
 ## [0.0.7]